cleanup imports, move config resolution to separate file (#222)
* cleanup imports, move config resolution to separate file

* config resolver tests

* add more tests
