+++ sgrep_lint/tests/python/eqeq.expected.remote.json
@@ -1,0 +1,1 @@
+{"results": [{"check_id": "eqeq-is-bad", "path": "tests/lint/stupid.py", "start": {"line": 3, "col": 12}, "end": {"line": 3, "col": 26}, "extra": {"message": "a+b == a+b is stupid", "metavars": {"$X": {"start": {"line": 3, "col": 12, "offset": 55}, "end": {"line": 3, "col": 17, "offset": 60}, "abstract_content": "a+b"}}}}, {"check_id": "eqeq-is-bad", "path": "tests/lint/stupid.py", "start": {"line": 8, "col": 5}, "end": {"line": 8, "col": 11}, "extra": {"message": "x == x is stupid", "metavars": {"$X": {"start": {"line": 8, "col": 5, "offset": 168}, "end": {"line": 8, "col": 6, "offset": 169}, "abstract_content": "x"}}}}, {"check_id": "eqeq-is-bad", "path": "tests/lint/stupid.py", "start": {"line": 12, "col": 12}, "end": {"line": 12, "col": 18}, "extra": {"message": "x == x is stupid", "metavars": {"$X": {"start": {"line": 12, "col": 12, "offset": 260}, "end": {"line": 12, "col": 13, "offset": 261}, "abstract_content": "x"}}}}]}
